Make import_backup_file_silently() create the course, to retain more course settings
The old way of creating the course first and then using import_backup_file_silently()
to overwrite it, only overwrites the course contents but doesn't preserve any of the
course settings except those that are copied over by hand later.

By passing in a 0 for $courseid, you force import_backup_file_silently() to create the
course, copying all the course settings over from the backup file. This is sort of an
undocumented feature of the function which doesn't work in standard Moodle because it
fails to create a context for the course. The associated change to onenine.patch fixes
that problem, though.