setup/patches/onenine.patch: Include course module data by default
The Moodle 2 export process uses Moodle's automated course backup settings
to decide what data to include and what to exclude. Ultimately the setting
for whether or not to include course activities and resources (which are
the main thing you'd want to convert!) derives from backup_sche_modules
in the initial Moodle 1.9 setup. And if you're creating one from scratch,
this defaults to "no". So, this patch changes it to say "yes" instead.