README.txt: Include config settings for php "variables_order" flag
Many of the arguments to the command line scripts are passed in via
environment variables, but they won't be sent to the script if
the "variables_order" flag in php.ini doesn't include "E". In some
distributions "E" is removed by default for security purposes, so it's
important to document this.